pvtLeechbot/pyproject.toml
[project]
name = "pvtleechbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.1",
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
